refactor: Used LazyModal for modals
