+++ src/components/utils/ZoomedImage.jsx
@@ -1,17 +1,17 @@
-"use client";
+import styles from "./ZoomedImage.module.css";
 import dynamic from "next/dynamic";
-import { useEffect, useState } from "react";
-const ModalZoomedImage = dynamic(() => import("../modals/ModalZoomedImage"), {
+const LazyModal = dynamic(() => import("../transitions/LazyModal"), {
   ssr: false,
 });
 
-export default function ZoomedImage(props) {
-  const { show } = props;
-  const [mounted, setMounted] = useState(show);
-
-  useEffect(() => {
-    if (show) setMounted(true);
-  }, [show]);
-
-  return mounted && <ModalZoomedImage {...props} />;
+export default function ZoomedImage({ show, onHide, src }) {
+  return (
+    <LazyModal show={show} onHide={onHide} centered>
+      <div className="modal-body p-0 p-relative">
+        <div className={styles.zoomedImageContainer}>
+          <img src={src} className={styles.zoomedImage} />
+        </div>
+      </div>
+    </LazyModal>
+  );
 }
